third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/custom-element-wrapper-gc-expected.txt

CONSOLE WARNING: line 14: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
Custom element wrappers in templates should be retained

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing major GC...
PASS customElement.expando is "present"
Testing minor GC...
PASS customElement.expando is "present"
PASS successfullyParsed is true

TEST COMPLETE
